fix(Backend): Fixed api to get rate for service and refactor implementation

+++ Backend/FishingBooker/FishingBooker/src/main/java/com/isa/FishingBooker/service/RevisionService.java
@@ -10,7 +10,7 @@ public interface RevisionService extends Service<Revision> {
 	
 	public int getBoatAppointmentRevisionsRate(Integer id);
 	public int getResortAppointmentRevisionsRate(Integer id);
-	public int getTutorServiceAppointmentRevisionsRate(Integer id);
+	public double getTutorServiceAppointmentRevisionsRate(Integer id);
 	public String makeTutorServiceRevision(TutorServiceAppointmentRevision revision);
 	public String makeBoatRevision(BoatAppointmentRevision revision);
 	public String makeResortRevision(ResortAppointmentRevision revision);
